Improve script structure

+++ src/appCataloga/root/usr/local/bin/appCataloga/appCataloga_stop.sh
@@ -1,24 +1,39 @@
 #!/bin/bash
-echo "All appCataloga services will be stopped"
+# This script stops all appCataloga services
 
-systemctl stop appCataloga_host_check
-echo "appCataloga_host_check stopped"
+splash_banner() {
+    # print full screen splash screen with the message received as argument
+
+    terminal_width=$(tput cols)
+
+    echo -e "\e[32m$(printf "%0.s~" $(seq 1 $terminal_width))\e[0m"
+    printf "\e[32m%*s\e[0m\n" $((($terminal_width + ${#1}) / 2)) "$1"
+    echo -e "\e[32m$(printf "%0.s~" $(seq 1 $terminal_width))\e[0m"
+}
+
+splash_banner "AppCataloga Service Stopper"
+
+read -p "All appCataloga services will be stopped. Do you want to continue? [y/N]" -n 1 -r
+echo
+
+if [[ ! $REPLY =~ ^[Yy]$ ]]; then
+    echo "Operation canceled"
+    exit 1
+fi
+
+services=("appCataloga.service" "appCataloga_file_bin_proces.service" "appCataloga_host_check.service" "appCataloga_pub_metadata.service")
+
+for i in "${services[@]}"; do
+    echo "$i Stop requested"
+    systemctl stop "$i"
+done
 
 # Create a list with all instances of appCataloga_file_bkp@* services
 # and stop them
 for i in $(systemctl list-units --full --all | grep appCataloga_file_bkp@ | awk '{print $1}'); do
+    echo "$i Stop requested"
     systemctl stop "$i"
-    echo "$i stopped"
 done
-
-systemctl stop appCataloga_file_bin_proces.service
-echo "appCataloga_file_bin_proces.service stopped"
-
-systemctl stop appCataloga_pub_metadata.service
-echo "appCataloga_pub_metadata.service stopped"
-
-systemctl stop appCataloga.service
-echo "appCataloga.service stopped"
 
 read -p "Remove log file? [y/N] " -n 1 -r
 echo
@@ -28,3 +43,6 @@ else
     rm -f /var/log/appCataloga.log
     echo "Log file removed"
 fi
+
+echo bye
+echo
